+++ config/test.exs
@@ -2,7 +2,7 @@ use Mix.Config
 
 config :k8s,
   discovery_provider: Mock.Discovery,
-  http_provider: Mock.HTTPProvider,
+  http_provider: K8s.Mock.DynamicHTTPProvider,
   clusters: %{
     test: %{
       conf: "test/support/kube-config.yaml"