+++ test/k8s/client/runner/base_test.exs
@@ -1,39 +1,48 @@
+# credo:disable-for-this-file
 defmodule K8s.Client.Runner.BaseTest do
-  # credo:disable-for-this-file
   use ExUnit.Case, async: true
   doctest K8s.Client.Runner.Base
 
   alias K8s.Client
   alias K8s.Client.Runner.Base
+  alias K8s.Mock.DynamicHTTPProvider
+  import K8s.Test.KubeHelper
+
+  defmodule HTTPMock do
+    @base_url "https://localhost:6443"
+    @namespaced_url @base_url <> "/api/v1/namespaces"
+    import K8s.Test.HTTPHelper
+
+    def request(:get, @namespaced_url, _, _, _), do: render(nil)
+
+    def request(:post, @namespaced_url, _, _, _), do: render(nil)
+
+    def request(:get, @namespaced_url <> "/test", _body, _headers, _opts) do
+      render(nil)
+    end
+  end
 
   setup do
+    DynamicHTTPProvider.register(self(), __MODULE__.HTTPMock)
     conf = K8s.Conf.from_file("test/support/kube-config.yaml")
     cluster = K8s.Cluster.register(:base_runner_test, conf)
 
     {:ok, cluster: cluster}
   end
 
-  def namespace_manifest() do
-    %{
-      "apiVersion" => "v1",
-      "metadata" => %{"name" => "test"},
-      "kind" => "Namespace"
-    }
-  end
-
   describe "run/3" do
     test "running an operation without an HTTP body", %{cluster: cluster} do
-      operation = Client.get(namespace_manifest())
+      operation = Client.get(make_namespace("test"))
       assert {:ok, _} = Base.run(operation, cluster)
     end
 
     test "running an operation with an HTTP body", %{cluster: cluster} do
-      operation = Client.create(namespace_manifest())
+      operation = Client.create(make_namespace("test"))
       assert {:ok, _} = Base.run(operation, cluster)
     end
 
     test "running an operation with options", %{cluster: cluster} do
-      operation = Client.get(namespace_manifest())
+      operation = Client.get(make_namespace("test"))
       opts = [params: %{"watch" => "true"}]
       assert {:ok, _} = Base.run(operation, cluster, opts)
     end
@@ -41,9 +50,9 @@ defmodule K8s.Client.Runner.BaseTest do
 
   describe "run/4" do
     test "running an operation with a custom HTTP body", %{cluster: cluster} do
-      operation = Client.create(namespace_manifest())
+      operation = Client.create(make_namespace("test"))
       labels = %{"env" => "test"}
-      body = put_in(namespace_manifest(), ["metadata", "labels"], labels)
+      body = put_in(make_namespace("test"), ["metadata", "labels"], labels)
 
       assert {:ok, _} = Base.run(operation, cluster, body)
     end
@@ -51,9 +60,9 @@ defmodule K8s.Client.Runner.BaseTest do
     test "running an operation with a custom HTTP body and options", %{
       cluster: cluster
     } do
-      operation = Client.create(namespace_manifest())
+      operation = Client.create(make_namespace("test"))
       labels = %{"env" => "test"}
-      body = put_in(namespace_manifest(), ["metadata", "labels"], labels)
+      body = put_in(make_namespace("test"), ["metadata", "labels"], labels)
       opts = [params: %{"watch" => "true"}]
       assert {:ok, _} = Base.run(operation, cluster, body, opts)
     end