+++ test/k8s/discover_test.exs
@@ -1,4 +1,0 @@
-defmodule K8s.DiscoveryTest do
-  use ExUnit.Case, async: true
-  doctest K8s.Discovery
-end