+++ test/k8s/group_test.exs
@@ -30,18 +30,18 @@ defmodule K8s.GroupTest do
 
   describe "find_resource_by_name/2" do
     test "finds a resource by atom name" do
-      {:ok, dep} = K8s.Group.find_resource_by_name(resources(), :deployment)
-      assert %{"kind" => "Deployment"} = dep
+      {:ok, deployment} = K8s.Group.find_resource_by_name(resources(), :deployment)
+      assert %{"kind" => "Deployment"} = deployment
     end
 
     test "finds a resource by plural name" do
-      {:ok, dep} = K8s.Group.find_resource_by_name(resources(), :deployments)
-      assert %{"kind" => "Deployment"} = dep
+      {:ok, deployment} = K8s.Group.find_resource_by_name(resources(), :deployments)
+      assert %{"kind" => "Deployment"} = deployment
     end
 
     test "finds a resource by kind name" do
-      {:ok, dep} = K8s.Group.find_resource_by_name(resources(), "Deployment")
-      assert %{"kind" => "Deployment"} = dep
+      {:ok, deployment} = K8s.Group.find_resource_by_name(resources(), "Deployment")
+      assert %{"kind" => "Deployment"} = deployment
     end
 
     test "returns an error when the resource is not supported" do