+++ test/test_helper.exs
@@ -1,1 +1,2 @@
+K8s.Mock.DynamicHTTPProvider.start_link(nil)
 ExUnit.start()